Cleaning Up, Fixing "server" related issues And New Features (#12)
Reworking the bootstrapper to focus on ease and simplicity of initial set-up. We've removed a lot of configuration options that prompt when setting up a new project, and we've set "sensible defaults" for everything else.

This commit also fixes the core `pytest-cookies` tests and adds a Github action to run the tests automatically on push. Finally, we've added an initial database migration for the 'core' Django app of generated projects.

Co-authored-by: Moussa Mokhtari <m_mokhtari@esi.dz>
Co-authored-by: Ed Romano <oudeis.metis@gmail.com>
Co-authored-by: William Huster <william@thinknimble.com>