Fix test settings so the built in `test_password_reset` passes (#296)
I've had to fix this on a couple projects. The built-in unit test test_password_reset fails locally because Django whitenoise CompressedManifestStaticFilesStorage expects logo.png to exist in the manifest. But that only works after collectstatic has been called, making this test flaky.

This PR adjusts the STATICFILES_STORAGE value in test_settings.py so that the whitenoise manifest check is bypassed and so test_password_reset will pass.