Remove jinja2-time from _extensions (#350)
According to cookiecutter/cookiecutter#1888, the jinja2-time extension is no longer required and can be removed. In more recent versions of cookiecutter, the time extension is loaded automatically. This should make it possible to run the bootstrapper with the pipx, which is recommended by cookiecutter's docs.

This PR removes the jinja2-time extension and updates the installation instructions in the README to use pipx.