Update the docker file for uv (#393)
## What this does

Updates Boto3 since the current version in the Bootsrapper does not work
with python 3.12
Moves the chat component to its own urls file 
Separates out the URLS for core
Adds sorting extension to RUFF

## How to test

Add user steps to achieve desired functionality for this feature.

## What this does

Add a description of what feature/page this PR intends to add or change.

## Checklist
- [ ] Todo 1
- [ ] Todo 2
- [ ] Todo 3


## How to test

Add user steps to achieve desired functionality for this feature.