remove docker container names and switch to system python (#394)
## What this does

@Jsicardi noticed that there was an issue with .venv being broken in the
container, the issue is that the .venv file from the host is being
copied over to the container. This is different than befoew when pipenv
created a python env in a different directory than the project directory
so this was not an issue, additionally we were using the pipfile.lock
without a container so this change does that.


The client had a set name for its container, this was causing a clash
because the container was the same for every project.
Having predictable names helps when running commands!

---------

Co-authored-by: Julian Sicardi <161532806+JSicardiNC@users.noreply.github.com>